requirements.txt
streamlit
pandas
plotly
joblib
streamlit-option-menu
scikit-learn
shap
matplotlib
numpys
pyarrow
xlsxwriter
google-genai